reduce MK4S leveling and nozzle cleaning temp for PC
during the bed leveling and nozzel cleaning step, PC has it's temp set a bit high and like to ooze out of the nozzle. This causes the print start to fail for me about 50-75% of the time due to nozzle cleaning failure.

Reducing the temp during this stage corrects the problem entirely as the filament does not ooze out of the nozzle.